% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logfile_existence.R
\name{check_logfile_existence}
\alias{check_logfile_existence}
\title{Check, if a logfile exists}
\usage{
check_logfile_existence(logfile_path = default_logfile_path())
}
\arguments{
\item{logfile_path}{path of the log file. Defaults to the global variable `LOGFILE_PATH`
if it exists, otherwise `"logfile/logfile.csv"`.}
}
\value{
a data frame
}
\description{
This function loads the logfile, if it exists. If it doesn't, it returns an empty log
with the expected columns.
}
\examples{
check_logfile_existence(tempfile(fileext = ".csv"))
}
\seealso{
\code{\link{generate_logfile_entry}} get the data to be logged
}
