% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(formula, range_from = -2 * pi, range_to = 2 * pi, step = 0.01)
}
\arguments{
\item{formula}{a list that contains formulas for transforming the x- and y-values.
The formulas can use the variables `x_i` and `y_i`.}

\item{range_from, range_to}{start and end of the base sequence. Default is `-2 * pi` to `2 * pi`.}

\item{step}{step size of the base sequence. Smaller steps give more points (the number of
points grows quadratically) and therefore denser images, but take longer.}
}
\value{
data frame
}
\description{
The generative images are based on values in a dataframe. This function creates a grid of
all combinations of `x_i` and `y_i` from `seq(range_from, range_to, by = step)` and
transforms it with a `formula`.
}
\examples{
formula <- list(
  x = quote(runif(1, -1, 1) * x_i^2 - sin(y_i^2)),
  y = quote(runif(1, -1, 1) * y_i^3 - cos(x_i^2))
)
df <- generate_data(formula, step = 0.1)
}
\seealso{
\code{\link{generate_plot}} the returned data frame is the input to generate the plot
}
