% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_logfile_entry.R
\name{generate_logfile_entry}
\alias{generate_logfile_entry}
\title{Append an entry to the logfile}
\usage{
generate_logfile_entry(
  logfile,
  formula,
  seed,
  file_name,
  params = list(),
  logfile_path = default_logfile_path()
)
}
\arguments{
\item{logfile}{the logfile, as returned by `check_logfile_existence()`}

\item{formula}{the formula list that is used}

\item{seed}{the specific seed used}

\item{file_name}{the file name of the image}

\item{params}{named list of the data and plot parameters (see `generate_img()`).
Missing values are filled with the package defaults.}

\item{logfile_path}{path of the log file. Defaults to the global variable `LOGFILE_PATH`
if it exists, otherwise `"logfile/logfile.csv"`.}
}
\value{
the updated log data frame (invisibly); the log file is written as a side effect
}
\description{
This function saves the information needed to recreate an image with random elements
(seed, formula and all data and plot parameters) and appends it to the logfile.
}
\examples{
path <- tempfile(fileext = ".csv")
formula <- list(x = quote(sin(x_i)), y = quote(cos(y_i)))
generate_logfile_entry(check_logfile_existence(path), formula, 42, "img.png",
                       logfile_path = path)
}
\seealso{
\code{\link{check_logfile_existence}} check, if there is a logfile
}
