% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed_from_logfile.R
\name{get_seed_from_logfile}
\alias{get_seed_from_logfile}
\title{Get one or several seeds from the logfile}
\usage{
get_seed_from_logfile(seed_to_recreate, logfile_path = default_logfile_path())
}
\arguments{
\item{seed_to_recreate}{the seed (or seeds) of the original image(s)}

\item{logfile_path}{path of the log file. Defaults to the global variable `LOGFILE_PATH`
if it exists, otherwise `"logfile/logfile.csv"`.}
}
\value{
a vector of the seeds that are found in the log file
}
\description{
Get one or several seeds from the logfile
}
\examples{
\dontrun{
get_seed_from_logfile(104)
}
}
\seealso{
\code{\link{get_formula_from_logfile}} get the formula for recreation

\code{\link{regenerate_img}} function to recreate an image
}
