% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenerate_img.R
\name{regenerate_img}
\alias{regenerate_img}
\title{Recreate an image}
\usage{
regenerate_img(
  seed_to_recreate,
  filetype = "png",
  polar = NULL,
  ...,
  img_path = default_img_path(),
  logfile_path = default_logfile_path()
)
}
\arguments{
\item{seed_to_recreate}{the seed of the original image}

\item{filetype}{set the file type for the image}

\item{polar}{logical, overrides the coordinate system stored in the log file}

\item{...}{further parameters overriding the logged ones, see `generate_img()`}

\item{img_path}{directory the image is saved in. Defaults to the global variable
`IMG_PATH` if it exists, otherwise `"img/everything/"`.}

\item{logfile_path}{path of the log file. Defaults to the global variable `LOGFILE_PATH`
if it exists, otherwise `"logfile/logfile.csv"`.}
}
\value{
the path of the saved image (invisibly)
}
\description{
This is the main function to recreate a previously generated image. It takes all relevant
information from the log file to get exactly the same result: by using the same formula and
the same seed you get the same random numbers. The plot parameters (polar, colors, alpha,
etc.) are taken from the log file as well, but can be overridden, e.g. to render the same
image in another color. Log entries written by older versions of the package do not contain
these parameters; the current defaults are used for them.
}
\examples{
\dontrun{
regenerate_img(104)
regenerate_img(104, color = "white", background_color = "black")
}
}
\seealso{
\code{\link{get_formula_from_logfile}} get the formula for recreation

\code{\link{get_seed_from_logfile}} get the seed, which is used as input
}
