% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_directories.R
\name{setup_directories}
\alias{setup_directories}
\title{Setup a directory structure for image creation}
\usage{
setup_directories(img_path, img_subdir, img_subdir2, logfile_dir)
}
\arguments{
\item{img_path}{the path, where the images should be saved}

\item{img_subdir}{one subdirectory}

\item{img_subdir2}{a second subdirectory}

\item{logfile_dir}{the log file directory}
}
\value{
the created directories (invisibly)
}
\description{
I use a certain directory structure for image creation. Every image is saved under "img/everything/". I use the directory "img/handpicked/" for a selection of good images. Moreover I create a directory for the log file.
You only need to call this function one time.
}
\examples{
\dontrun{
setup_directories("img/", "everything/", "handpicked/", "logfile/")
}
}
